Изменение бизнес-логики. Доработка компонентов. Правка багов.

+++ src/main/java/com/gazprom/system/payload/UserRequest.java
@@ -15,5 +15,5 @@ public class UserRequest {
     private String middleName;
     private String email;
     private Long departmentId;
-    private Long role;
+    private int role;
 }
